Reestructurando Problema funcionalidad 4
