+++ src/main/java/com/clinicasalud/Clinica/Salud/ClinicaSaludApplication.java
@@ -1,9 +1,9 @@
 package com.clinicasalud.Clinica.Salud;
 
 import com.clinicasalud.Clinica.Salud.model.cita.CitaService;
+import com.clinicasalud.Clinica.Salud.model.horariomedico.HorarioMedicoJpaRepository;
 import com.clinicasalud.Clinica.Salud.model.medico.EspecialidadService;
 import com.clinicasalud.Clinica.Salud.model.medico.MedicoService;
-import com.clinicasalud.Clinica.Salud.model.medico.MedicoRepository;
 import com.clinicasalud.Clinica.Salud.model.paciente.PacienteRepository;
 import com.clinicasalud.Clinica.Salud.model.paciente.PacienteService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,33 +12,16 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 
-@SpringBootApplication
+@SpringBootApplication()
 public class ClinicaSaludApplication implements CommandLineRunner{
 
 	public static void main(String[] args) {
 		SpringApplication.run(ClinicaSaludApplication.class, args);
 	}
-
 	@Autowired
-	private PacienteRepository pacienteRepository;
-
-	@Autowired
-	private PacienteService pacienteService;
-
-	@Autowired
-	private EspecialidadService especialidadService;
-
-	@Autowired
-	private MedicoService medicoService;
-
-	@Autowired
-	private CitaService citaService;
-
-	//Ejecucion en Consola
+	private Principal principal;
 	@Override
 	public void run(String... args) throws Exception {
-
-		Principal principal = new Principal(pacienteRepository, pacienteService,especialidadService, medicoService, citaService);
 		principal.menu();
 	}
 }
