+++ src/main/java/com/clinicasalud/Clinica/Salud/model/apoyodiagnostico/ApoyoDiagnosticoJpaRepository.java
@@ -2,5 +2,5 @@ package com.clinicasalud.Clinica.Salud.model.apoyodiagnostico;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApoyoDiagnosticoJpaService extends JpaRepository<ApoyoDiagnostico,Long> {
+public interface ApoyoDiagnosticoJpaRepository extends JpaRepository<ApoyoDiagnostico,Long> {
 }
