+++ src/main/java/com/clinicasalud/Clinica/Salud/model/diaatencion/DiaAtencionJpaRepository.java
@@ -2,5 +2,5 @@ package com.clinicasalud.Clinica.Salud.model.diaatencion;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaAtencionRepository extends JpaRepository<DiaAtencion,Long> {
+public interface DiaAtencionJpaRepository extends JpaRepository<DiaAtencion,Long> {
 }
