+++ src/main/java/com/clinicasalud/Clinica/Salud/model/diagnostico/DiagnosticoJpaRepository.java
@@ -1,7 +1,6 @@
 package com.clinicasalud.Clinica.Salud.model.diagnostico;
 
-import jdk.jshell.Diag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiagnosticoRepository extends JpaRepository<Diagnostico,Long> {
+public interface DiagnosticoJpaRepository extends JpaRepository<Diagnostico,Long> {
 }
