+++ src/main/java/com/clinicasalud/Clinica/Salud/model/horariodia/HorarioDiaJpaRepository.java
@@ -2,5 +2,5 @@ package com.clinicasalud.Clinica.Salud.model.horariodia;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HorarioDiaRepository extends JpaRepository<HorarioDia,Long> {
+public interface HorarioDiaJpaRepository extends JpaRepository<HorarioDia,Long> {
 }
