+++ src/main/java/com/clinicasalud/Clinica/Salud/model/horariomedico/HorarioMedicoJpaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HorarioMedicoJpaService extends JpaRepository<HorarioMedico, Long> {
+public interface HorarioMedicoJpaRepository extends JpaRepository<HorarioMedico, Long> {
 }
