--- src/main/java/com/clinicasalud/Clinica/Salud/model/medico/MedicoJpaService.java
@@ -1,6 +1,0 @@
-package com.clinicasalud.Clinica.Salud.model.medico;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MedicoJpaService extends JpaRepository<Medico, Long> {
-}
