+++ src/main/java/com/clinicasalud/Clinica/Salud/model/tratamiento/TratamientoJpaRepository.java
@@ -2,5 +2,5 @@ package com.clinicasalud.Clinica.Salud.model.tratamiento;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TratamientoJpaService extends JpaRepository<Tratamiento,Long> {
+public interface TratamientoJpaRepository extends JpaRepository<Tratamiento,Long> {
 }
