+++ src/main/java/com/clinicasalud/Clinica/Salud/model/usuario/UsuarioJpaRepository.java
@@ -2,5 +2,5 @@ package com.clinicasalud.Clinica.Salud.model.usuario;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioJpaService extends JpaRepository<Usuario,Long> {
+public interface UsuarioJpaRepository extends JpaRepository<Usuario,Long> {
 }
